move tests to appropriate locations and fix tests that weren't quite right
