+++ src/NServiceBus.AcceptanceTests/Basic/When_handling_current_message_later.cs
@@ -101,7 +101,6 @@
             }
         }
 
-        
         public class SomeMessage : IMessage
         {
             public Guid Id { get; set; }
