+++ src/NServiceBus.AcceptanceTests/Core/Pipeline/When_aborting_the_behavior_chain.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Basic
+﻿namespace NServiceBus.AcceptanceTests.Core.Pipeline
 {
     using System.Threading.Tasks;
     using AcceptanceTesting;
