+++ src/NServiceBus.AcceptanceTests/Core/Pipeline/When_replacing_behavior.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Pipeline
+﻿namespace NServiceBus.AcceptanceTests.Core.Pipeline
 {
     using System;
     using System.Threading.Tasks;
