+++ src/NServiceBus.AcceptanceTests/NServiceBus.AcceptanceTests.csproj
@@ -55,8 +55,8 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Basic\When_depending_on_untyped_feature.cs" />
-    <Compile Include="Basic\When_extending_behavior_context.cs" />
+    <Compile Include="Core\Feature\When_depending_on_untyped_feature.cs" />
+    <Compile Include="Core\Pipeline\When_extending_behavior_context.cs" />
     <Compile Include="Basic\When_receiving_unobtrusive_message_without_handler.cs" />
     <Compile Include="Config\When_multiple_configuration_providers.cs" />
     <Compile Include="Core\CoreTestSuiteConstraints.cs" />
@@ -116,7 +116,7 @@
     <Compile Include="ScenarioDescriptors\TypeScanner.cs" />
     <Compile Include="Serialization\When_configuring_custom_xml_namespace.cs" />
     <Compile Include="Serialization\When_registering_additional_deserializers.cs" />
-    <Compile Include="Basic\When_no_content_type.cs" />
+    <Compile Include="Serialization\When_no_content_type.cs" />
     <Compile Include="Basic\When_registering_handlers_explicitly.cs" />
     <Compile Include="Causation\When_a_message_is_faulted.cs" />
     <Compile Include="Causation\When_a_message_is_audited.cs" />
@@ -138,13 +138,13 @@
     <Compile Include="Core\FakeTransport\FakeTransportInfrastructure.cs" />
     <Compile Include="Core\FakeTransport\FakeTransportSetingsExtensions.cs" />
     <Compile Include="Core\FakeTransport\ProcessingOptimizations\When_using_concurrency_limit.cs" />
-    <Compile Include="Basic\When_depending_on_typed_feature.cs" />
+    <Compile Include="Core\Feature\When_depending_on_typed_feature.cs" />
     <Compile Include="Hosting\When_overriding_input_queue_name.cs" />
     <Compile Include="Hosting\When_a_message_is_faulted.cs" />
     <Compile Include="MessageId\When_message_has_empty_id_header.cs" />
     <Compile Include="MessageId\When_message_has_no_id_header.cs" />
     <Compile Include="CriticalError\When_raising_critical_error.cs" />
-    <Compile Include="Basic\When_deferring_to_non_local.cs" />
+    <Compile Include="DelayedDelivery\When_deferring_to_non_local.cs" />
     <Compile Include="Core\DelayedDelivery\TimeoutManager\When_TimeoutManager_is_disabled.cs" />
     <Compile Include="Mutators\When_incoming_mutator_changes_message_type.cs" />
     <Compile Include="Mutators\When_mutating.cs" />
@@ -163,8 +163,8 @@
     <Compile Include="Core\Persistence\When_a_persistence_does_not_support_saga.cs" />
     <Compile Include="Core\Persistence\When_a_persistence_does_not_support_subscriptions.cs" />
     <Compile Include="Core\Persistence\When_a_persistence_does_not_support_timeouts.cs" />
-    <Compile Include="Pipeline\When_using_per_uow_component_in_the_pipeline.cs" />
-    <Compile Include="Pipeline\When_replacing_behavior.cs" />
+    <Compile Include="Core\Pipeline\When_using_per_uow_component_in_the_pipeline.cs" />
+    <Compile Include="Core\Pipeline\When_replacing_behavior.cs" />
     <Compile Include="Core\Persistence\When_a_persistence_does_not_provide_ISynchronizationContext.cs" />
     <Compile Include="Recoverability\Retries\When_message_is_deferred_by_delayed_retries_using_dtc.cs" />
     <Compile Include="Recoverability\Retries\When_message_fails_retries.cs" />
@@ -190,9 +190,9 @@
     <Compile Include="Audit\When_a_replymessage_is_audited.cs" />
     <Compile Include="Performance\When_message_is_audited.cs" />
     <Compile Include="Basic\When_receiving_with_catch_all_handlers_registered.cs" />
-    <Compile Include="Basic\When_handling_message_with_several_messagehandlers.cs" />
+    <Compile Include="Core\Pipeline\When_handling_message_with_several_messagehandlers.cs" />
     <Compile Include="Serialization\When_registering_custom_serializer.cs" />
-    <Compile Include="ApiExtension\When_extending_sendoptions.cs" />
+    <Compile Include="Core\Pipeline\When_extending_sendoptions.cs" />
     <Compile Include="Basic\When_sending_with_conventions.cs" />
     <Compile Include="Basic\When_sending_from_a_send_only.cs" />
     <Compile Include="Basic\When_using_a_greedy_convention.cs" />
@@ -237,7 +237,7 @@
     <Compile Include="Routing\When_publishing_an_interface.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\When_publishing_from_sendonly.cs" />
     <Compile Include="Routing\When_publishing_an_event_implementing_two_unrelated_interfaces.cs" />
-    <Compile Include="ApiExtension\When_extending_the_publish_api.cs" />
+    <Compile Include="Core\Pipeline\When_extending_the_publish_api.cs" />
     <Compile Include="Routing\When_replying_to_message.cs" />
     <Compile Include="RunDescriptorExtensions.cs" />
     <Compile Include="Sagas\When_forgetting_to_set_a_corr_property.cs" />
@@ -295,9 +295,9 @@
     <Compile Include="Audit\When_auditing.cs" />
     <Compile Include="Audit\When_a_message_is_audited.cs" />
     <Compile Include="Correlation\When_sending_with_no_correlation_id.cs" />
-    <Compile Include="Basic\When_aborting_the_behavior_chain.cs" />
+    <Compile Include="Core\Pipeline\When_aborting_the_behavior_chain.cs" />
     <Compile Include="Basic\When_handling_current_message_later.cs" />
-    <Compile Include="Basic\When_multiple_mappings_exists.cs" />
+    <Compile Include="Routing\When_multiple_mappings_exists.cs" />
     <Compile Include="Correlation\When_using_a_custom_correlation_id.cs" />
     <Compile Include="Config\When_only_abstract_config_override_is_found.cs" />
     <Compile Include="Core\Encryption\When_using_encryption_with_custom_service.cs" />
