+++ src/NServiceBus.AcceptanceTests/Routing/When_multiple_mappings_exists.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Basic
+﻿namespace NServiceBus.AcceptanceTests.Routing
 {
     using System.Threading.Tasks;
     using AcceptanceTesting;
