+++ src/NServiceBus.AcceptanceTests/Serialization/When_no_content_type.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Basic
+﻿namespace NServiceBus.AcceptanceTests.Serialization
 {
     using System.Threading.Tasks;
     using AcceptanceTesting;
